refactor: restructure project to feature-based folder architecture
