+++ src/features/ProductList/components/Dropdown.module.css
@@ -66,7 +66,7 @@
   }
 
   .dropdownButton > img {
-    content: url(../assets/ic_arrow_down.svg);
+    content: url(assets/images/ic_arrow_down.svg);
   }
 
   .dropdownButton > span {
