+++ src/index.js
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
-import "./styles/reset.css";
-import "./styles/variables.css";
+import "assets/styles/reset.css";
+import "assets/styles/variables.css";
 import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
